pyproject.toml
[project]
name = "basic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.0",
    "openai>=2.28.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.0.0",
]
